finish other section

+++ resources/item.py
@@ -1,9 +1,8 @@
-import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
-
+from flask_jwt_extended import jwt_required, get_jwt
 from db import db
 from models import ItemModel
 from schema import ItemSchema,ItemUpdateSchema
@@ -13,6 +12,8 @@ blp = Blueprint("items", __name__, description="Operations on items")
 
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):
+    
+    @jwt_required()
     @blp.response(200,ItemSchema)
     def get(self, item_id):
        item = ItemModel.query.get_or_404(item_id)
@@ -33,8 +34,12 @@ class Item(MethodView):
         
         return item
     
-    
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_Admin"):
+            abort(403 , message="unauthorized")
+            
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -44,10 +49,14 @@ class Item(MethodView):
 
 @blp.route("/item")
 class itemList(MethodView):
+    
+    @jwt_required()
     @blp.response(200,ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201,ItemSchema)
     def post(self, item_data):
